Lifespan moves to middleware. TestClient can be used as a context manager.
