+++ starlette/applications.py
@@ -2,9 +2,9 @@ import typing
 
 from starlette.datastructures import URL, URLPath
 from starlette.exceptions import ExceptionMiddleware
-from starlette.lifespan import LifespanMiddleware
 from starlette.middleware.base import BaseHTTPMiddleware
 from starlette.middleware.errors import ServerErrorMiddleware
+from starlette.middleware.lifespan import LifespanMiddleware
 from starlette.routing import BaseRoute, Router
 from starlette.schemas import BaseSchemaGenerator
 from starlette.types import ASGIApp, ASGIInstance, Scope
